refactor(progression): separate profile rendering concerns
